split up "main_methods" into service modules
