+++ src/api/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service helpers used by API routes and orchestration."""
